% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_swica_cont}
\alias{scale_color_swica_cont}
\title{SWICA color scale functions for continuous variables}
\usage{
scale_color_swica_cont(name)
}
\description{
SWICA color scale functions for continuous variables
}
