% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_swica_discrete}
\alias{scale_color_swica_discrete}
\title{SWICA color scale functions for discrete variables}
\usage{
scale_color_swica_discrete(name)
}
\description{
SWICA color scale functions for discrete variables
}
