% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_colour_swica_cont}
\alias{scale_colour_swica_cont}
\title{SWICA colour scale functions for continuous variables}
\usage{
scale_colour_swica_cont(name)
}
\description{
SWICA colour scale functions for continuous variables
}
