% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_colour_swica_discrete}
\alias{scale_colour_swica_discrete}
\title{SWICA colour scale functions for discrete variables}
\usage{
scale_colour_swica_discrete(name)
}
\description{
SWICA colour scale functions for discrete variables
}
