% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_swica_discrete}
\alias{scale_fill_swica_discrete}
\title{SWICA fill functions for discrete values}
\usage{
scale_fill_swica_discrete(name)
}
\description{
SWICA fill functions for discrete values
}
