% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{swica_colours}
\alias{swica_colours}
\title{SWICA color palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
swica_colours
}
\description{
Provides ggplot themes and color palettes for SWICA
}
\keyword{datasets}
